Added 'basic' information to make Login Button work Important parts here are: Controllers/Account Models/ViewModels/LoginViewModel Views/Account/Login
These three additions now work together so that it's possible to click on Login and route to account/login
Will screenshot in discord and amend basic information soon.

Note that currently all that happens is we can click and go to the login page, login information is NOT currently registered from within the page. This and adding the register user page will be next ideas.